requirements.txt
Django==5.1.3
faiss-cpu
langchain-community
langchain-openai
openai
python-dotenv==1.0.1
tiktoken
pypdf
djangorestframework
PyPDF2
python-dotenv==1.0.1
gunicorn
psycopg[binary]
